changes requested in the code review
